+++ rlwe/src/lib.rs
@@ -1,3 +1,6 @@
 pub mod elem;
-pub mod encryption;
+pub mod elem_grlwe;
+pub mod elem_rgsw;
+pub mod elem_rlwe;
 pub mod keys;
+mod utils;
